+++ src/languages/de.ts
@@ -1,0 +1,4 @@
+export const de = {
+  wotd: "Das Wort des Tages ist:",
+  info: "Weitere Informationen finden Sie unter:"
+}