+++ src/languages/en.ts
@@ -1,0 +1,4 @@
+export const en = {
+  wotd: "The word of the day is:",
+  info: "You can find more information at:"
+}