+++ src/languages/es.ts
@@ -1,0 +1,4 @@
+export const es = {
+  wotd: "La palabra del día es:",
+  info: "Puede encontrar más información en:"
+}