+++ src/languages/fr.ts
@@ -1,0 +1,4 @@
+export const fr = {
+  wotd: "Le mot du jour est:",
+  info: "Vous trouverez de plus amples informations à l'adresse suivante:"
+}