+++ src/languages/pt.ts
@@ -1,0 +1,4 @@
+export const pt = {
+  wotd: "A palavra do dia é:",
+  info: "Pode encontrar mais informações em:"
+}