+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * **Fixed:** Print error in main function catch
 * **Fixed:** Missing color reset at the end of a printed line
 * **Update:** Build documentation
+* **Added:** Multi-language support with: Spanish, English, French, German, Portuguese
 
 ## [2.0.0] - 2020.05.23